refactor(core): typing improvements by using `isNgModule`.
This commit add the usage of `isNgModule` as type guard to prevent type assertions.
