+++ packages/core/src/render3/util/module_utils.ts
@@ -8,7 +8,7 @@
 
 import {ModuleWithProviders} from '../../di/interface/provider';
 import {Type} from '../../interface/type';
-import {NgModuleDef} from '../../metadata/ng_module_def';
+import {NgModuleType} from '../../metadata/ng_module_def';
 import {getComponentDef, getDirectiveDef, getNgModuleDef, getPipeDef} from '../definition';
 import {ComponentType, DirectiveType, PipeType} from '../interfaces/definition';
 
@@ -16,7 +16,7 @@ export function isModuleWithProviders(value: any): value is ModuleWithProviders<
   return (value as {ngModule?: any}).ngModule !== undefined;
 }
 
-export function isNgModule<T>(value: Type<T>): value is Type<T>&{ɵmod: NgModuleDef<T>} {
+export function isNgModule<T>(value: Type<T>): value is NgModuleType {
   return !!getNgModuleDef(value);
 }
